# docs/swagger.yaml
---
swagger: "2.0"

info:
  description: "Enternot is a home security system powered by the Raspberry PI."
  version: "0.0.1"
  title: "Enternot PI"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"

basePath: "/"

schemes:
- "http"

paths:
  /status:
    get:
      tags:
        - "server status"
      summary: "Test if the server is up and running"
      description: "Aliases for this are: /index and /"
      produces:
        - "application/json"
      responses:
          200:
            description: "OK"
            schema:
              $ref: "#/definitions/Status"
          401:
            description: "Unauthorized. Provide Basic Authentication header."
  /camera/stream:
    get:
      tags:
        - "camera"
      summary: "Live stream of the webcam"
      produces:
      - "multipart/x-mixed-replace"
      responses:
          200:
            description: "OK"
          401:
            description: "Unauthorized. Provide Basic Authentication header."
  /camera/position:
    post:
      tags:
        - "camera"
      summary: "Move the camera by the specified angles around an axis."
      consumes:
        - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The amount that the viewing angle for the camera for the X and Y axis should be changed by."
        required: true
        schema:
          $ref: "#/definitions/Angle"
      responses:
        200:
          description: "OK"
        400:
          description: "Bad request. Request body not valid json."
        401:
          description: "Unauthorized. Provide Basic Authentication header."
  /location:
    post:
      tags:
        - "notifications"
      summary: "Enable / Disable push notifications based on distance to the pi"
      description: "If user is more than 500 meters away from the pi, push notifications will be enabled, otherwise disabled."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Location of the user"
        required: true
        schema:
          $ref: "#/definitions/Location"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/LocationResponse"
        400:
          description: "Bad request. Request body not valid json."
        401:
          description: "Unauthorized. Provide Basic Authentication header."
  /siren/start:
    post:
      tags:
        - "audio playback"
      summary: "Start a playback of the siren sound for 3 minutes"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Message"
        401:
          description: "Unauthorized. Provide Basic Authentication header."
  /siren/stop:
    post:
      tags:
        - "audio playback"
      summary: "Stop the siren playback if it is running at the moment."
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Message"
        401:
          description: "Unauthorized. Provide Basic Authentication header."      

definitions:
  Message:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Some message from the server"
        example: "some message"

  Status:
    type: "object"
    properties:
      status:
        type: "string"
        description: "Message from the server"
        example: "Enternot is up and running!"
      server_time:
        type: "string"
        description: "Current server time"
        example: "Wed, 13 Jun 2018 19:13:34 GMT"
  Location:
    type: "object"
    properties:
      location:
        type: "object"
        properties:
          longitude:
            type: "number"
            example: 14.320049
          latitude:
            type: "number"
            example: 48.338040
  LocationResponse:
    type: "object"
    properties:
      notifications:
        type: "boolean"
        description: "Whether or not notifications are turned on"
        example: "true"
      distance:
        type: "number"
        description: "Distance to the pi, in meters"
        example: 446.34
  Angle:
    type: "object"
    properties:
      x_angle:
        type: "number"
        description: "X angle (left/right) in degrees, from -180 to 180"
        example: 12.5
      y_angle:
        type: "number"
        description: "Y angle (left/right) in degrees, from -180 to 180"
        example: -23.5